Autoloader patch.
This commit improves the package's autoloader by providing a better way to
arbitrarily add/remove paths without increasing code complexity and without
adding unnecessary code duplication to the autoloader. I would, however,
strongly recommend for the package to NOT declare additional autoloaders to
every single PHP file in the entire package in the future, and to consider
further refactoring. Declaring autoloaders in every single PHP file in the
entire package adds unnecessary complexity to the package. A better
approach would be to handle autoloading via a single, specific file which
can then be loaded as necessary by any other PHP files which need the
autoloader.